fix upload route styles
