+++ src/components/InputField/index.js
@@ -9,7 +9,13 @@ export const InputField = ({ imageURL, caption, onChange }) => {
   }, [caption]);
 
   return (
-    <div className="mx-auto my-5 h-20 rounded-lg border border-solid border-c700 flex items-center p-4 z-10 ">
+    <div
+      className="mx-auto mb-5 rounded-lg flex items-center p-2 z-10 "
+      style={{
+        boxShadow:
+          '0px 0px 1px rgba(0, 0, 0, 0.039), 0px 0.5px 1.5px rgba(0, 0, 0, 0.19)'
+      }}
+    >
       <img src={imageURL} alt="" className="w-8 h-8 rounded-full " />
       <textarea
         name="postText"
