+++ src/pages/post.jsx
@@ -2,8 +2,9 @@ import React, { useEffect, useContext } from 'react';
 import { UserContext } from '../context/userContext';
 import { useHistory } from 'react-router-dom';
 import { UploadSection } from '../components/UploadSection';
+import { Heading } from '../components/Heading';
 
-export const Post = ({ postDate }) => {
+export const Post = () => {
   const { user } = useContext(UserContext);
   const history = useHistory();
 
@@ -14,8 +15,19 @@ export const Post = ({ postDate }) => {
   });
 
   return (
-    <div className="h-full container ">
-      <div className="grid grid-cols-1 ">
+    <div className="bg-c900 py-3">
+      <div className="container">
+        <Heading
+          as="p"
+          fontSize="lg"
+          fontWeight="medium"
+          lineHeight="normal"
+          textAlign="left"
+          textColor="lightblack"
+          className="my-2 inline-block"
+        >
+          Post Something
+        </Heading>
         <UploadSection userImage={user.photoURL} />
       </div>
     </div>
